null check done

+++ salesyncer-server/business-service/src/database/repositories/activities-repo.ts
@@ -15,7 +15,7 @@ export const qGetActivitiesData = async () => {
       })
       .sort({ _id: -1 });
   } catch (error) {
-     logger.error(error);
+    logger.error(error);
   }
 };
 
@@ -23,12 +23,16 @@ export const qGetActivitiesData = async () => {
 
 export const qCreateActivityData = async (newActivityData: object) => {
   try {
-    const newActivity = new Activity(newActivityData);
+    if (newActivityData) {
+      const newActivity = new Activity(newActivityData);
 
-    const addActivityToDB = await newActivity.save();
-    return addActivityToDB;
+      const addActivityToDB = await newActivity.save();
+      return addActivityToDB;
+    } else {
+      logger.info(`Unable to write to db. Activity Data missing`);
+    }
   } catch (error) {
-     logger.error(error);
+    logger.error(error);
   }
 };
 
@@ -36,14 +40,18 @@ export const qCreateActivityData = async (newActivityData: object) => {
 
 export const qGetActivityDataById = async (_id: string) => {
   try {
-    return await Activity.findById(_id).populate({
-      path: "lead",
-      populate: {
-        path: "client",
-      },
-    });
+    if (_id) {
+      return await Activity.findById(_id).populate({
+        path: "lead",
+        populate: {
+          path: "client",
+        },
+      });
+    } else {
+      logger.info(`Unable to write to db. Data missing`);
+    }
   } catch (error) {
-     logger.error(error);
+    logger.error(error);
   }
 };
 
@@ -54,12 +62,16 @@ export const qUpdateActivityDataById = async (
   newActivityData: any
 ) => {
   try {
-    const updateOperation = {
-      $set: newActivityData,
-    };
-    const response = await Activity.findByIdAndUpdate(_id, updateOperation);
+    if (_id && newActivityData) {
+      const updateOperation = {
+        $set: newActivityData,
+      };
+      const response = await Activity.findByIdAndUpdate(_id, updateOperation);
 
-    return response;
+      return response;
+    } else {
+      logger.info(`Unable to write to db. Data missing`);
+    }
   } catch (error) {}
 };
 
@@ -67,10 +79,11 @@ export const qUpdateActivityDataById = async (
 
 export const qGetActivityCount = async () => {
   try {
-    const counterData: any = await BusinessCounter.findOne();
+    const counterData = await BusinessCounter.findOne();
+    if(counterData!==null)
     return counterData.activityCounter;
   } catch (error) {
-     logger.error(error);
+    logger.error(error);
   }
 };
 
@@ -79,21 +92,25 @@ export const qGetActivityCount = async () => {
 
 export const qIncActivityCount = async () => {
   try {
-    const updateCounterData: any = await BusinessCounter.findOneAndUpdate({
+    const updateCounterData = await BusinessCounter.findOneAndUpdate({
       $inc: { activityCounter: 1 },
     });
     return updateCounterData;
   } catch (error) {
-     logger.error(error);
+    logger.error(error);
   }
 };
 
 ////==============================================
 export const qDeleteActivityDataById = async (_id: string) => {
   try {
-    return await Activity.findByIdAndRemove(_id);
+    if (_id !==null) {
+      return await Activity.findByIdAndRemove(_id);
+    } else {
+      logger.info(`Unable to write to db. Data missing`);
+    }
   } catch (error) {
-     logger.error(error);
+    logger.error(error);
   }
 };
 
@@ -104,7 +121,7 @@ export const qGetActivityTypesData = async () => {
   try {
     return await ActivityType.find({});
   } catch (error) {
-     logger.error(error);
+    logger.error(error);
   }
 };
 
@@ -112,12 +129,16 @@ export const qGetActivityTypesData = async () => {
 
 export const qCreateActivityTypeData = async (newActivityTypeData: object) => {
   try {
-    const newActivityType = new ActivityType(newActivityTypeData);
+    if (newActivityTypeData) {
+      const newActivityType = new ActivityType(newActivityTypeData);
 
-    const addActivityTypeToDB = await newActivityType.save();
-    return addActivityTypeToDB;
+      const addActivityTypeToDB = await newActivityType.save();
+      return addActivityTypeToDB;
+    } else {
+      logger.info(`Unable to write to db. Data missing`);
+    }
   } catch (error) {
-     logger.error(error);
+    logger.error(error);
   }
 };
 
@@ -128,12 +149,19 @@ export const qUpdateActivityTypeDataById = async (
   newActivityTypeData: any
 ) => {
   try {
-    const updateOperation = {
-      $set: newActivityTypeData,
-    };
-    const response = await ActivityType.findByIdAndUpdate(_id, updateOperation);
+    if (_id && newActivityTypeData) {
+      const updateOperation = {
+        $set: newActivityTypeData,
+      };
+      const response = await ActivityType.findByIdAndUpdate(
+        _id,
+        updateOperation
+      );
 
-    return response;
+      return response;
+    } else {
+      logger.info(`Unable to write to db. Data missing`);
+    }
   } catch (error) {}
 };
 
@@ -141,9 +169,15 @@ export const qUpdateActivityTypeDataById = async (
 
 export const qDeleteActivityTypeDataById = async (_id: string) => {
   try {
-    return await ActivityType.findByIdAndRemove(_id);
+
+    if(_id){
+
+      return await ActivityType.findByIdAndRemove(_id);
+    } else {
+      logger.info(`Unable to write to db. Data missing`);
+    }
   } catch (error) {
-     logger.error(error);
+    logger.error(error);
   }
 };
 
@@ -154,18 +188,24 @@ export const qIsTimeCollisionExists = async (
   scheduledTimeInput: Date
 ) => {
   try {
-    return await Activity.find({
-      $and: [
-        { owner: ownerInput },
-        {
-          scheduledTime: {
-            $gte:  new Date(scheduledTimeInput).getTime() - 10 * 60 * 1000,
-            $lte: new Date(scheduledTimeInput).getTime() + 10 * 60 * 1000,
+    if(ownerInput && scheduledTimeInput){
+
+      
+      return await Activity.find({
+        $and: [
+          { owner: ownerInput },
+          {
+            scheduledTime: {
+              $gte: new Date(scheduledTimeInput).getTime() - 10 * 60 * 1000,
+              $lte: new Date(scheduledTimeInput).getTime() + 10 * 60 * 1000,
+            },
           },
-        },
-      ],
-    });
+        ],
+      });
+    } else {
+      logger.info(`Unable to write to db. Data missing`);
+    }
   } catch (error) {
-     logger.error(error);
+    logger.error(error);
   }
 };
